+++ packages/adhere-ui-comment/lib/index.js
@@ -1,0 +1,2 @@
+"use strict";require("core-js/modules/es.object.define-property.js"),Object.defineProperty(exports,"__esModule",{value:!0});var tslib_1=require("tslib"),comment_1=tslib_1.__importDefault(require("./comment"));exports.default=comment_1.default;
+//# sourceMappingURL=index.js.map