+++ packages/adhere-ui-comment/lib/types.js
@@ -1,0 +1,2 @@
+"use strict";require("core-js/modules/es.object.define-property.js"),Object.defineProperty(exports,"__esModule",{value:!0});
+//# sourceMappingURL=types.js.map