+++ packages/adhere-ui-flexlayout/lib/types.d.ts
@@ -1,18 +1,22 @@
 import React from 'react';
 export interface IFlexLayoutProps {
     direction: 'vertical' | 'horizontal';
-    className?: '';
+    className?: string;
     style?: React.CSSProperties;
+    children?: React.ReactNode | null;
 }
 export interface IFixedProps {
-    className?: '';
+    className?: string;
+    fit?: boolean;
     style?: React.CSSProperties;
+    children?: React.ReactNode | null;
 }
 export interface IAutoProps {
     autoFixed?: boolean;
     fit?: boolean;
-    className?: '';
+    className?: string;
     style?: React.CSSProperties;
+    children?: React.ReactNode | null;
 }
 export interface IVerticalFlexLayoutProps {
     className?: string;
@@ -42,6 +46,7 @@ export interface IVerticalFlexLayoutProps {
     leftProps?: object;
     mainProps?: object;
     mainAutoWrapProps?: object;
+    children?: any;
 }
 export interface IHorizontalFlexLayoutProps {
     className?: string;
@@ -71,6 +76,7 @@ export interface IHorizontalFlexLayoutProps {
     leftProps?: object;
     mainProps?: object;
     mainAutoWrapProps?: object;
+    children?: any;
 }
 export interface IToolBarLayoutProps {
     className?: string;