+++ packages/adhere-ui-flexlayout/src/fixed.tsx
@@ -4,7 +4,6 @@ import classNames from 'classnames';
 
 import { IFixedProps } from './types';
 
-// @ts-ignore
 import { FlexContext } from './context';
 
 const selectorPrefix = 'adhere-ui-flexlayout-fixed';
@@ -18,8 +17,6 @@ class Fixed extends React.Component<IFixedProps, any> {
   static defaultProps: any;
   static propTypes: any;
 
-  // @ts-ignore
-  private props: IFixedProps | undefined;
   private direction: string | undefined;
   private ref: RefObject<HTMLDivElement> | null | undefined;
 
@@ -36,7 +33,6 @@ class Fixed extends React.Component<IFixedProps, any> {
   renderInner({ direction }) {
     this.direction = direction;
 
-    // @ts-ignore
     const { children, className, style, fit } = this.props;
 
     return (
@@ -46,7 +42,7 @@ class Fixed extends React.Component<IFixedProps, any> {
         className={classNames(
           selectorPrefix,
           `${fit ? selectorPrefix + '-fit' : ''}`,
-          className.split(/\s+/),
+          (className || '').split(/\s+/),
         )}
       >
         {children}
@@ -55,7 +51,6 @@ class Fixed extends React.Component<IFixedProps, any> {
   }
 
   render() {
-    // @ts-ignore
     return <FlexContext.Consumer>{this.renderInner}</FlexContext.Consumer>;
   }
 }
@@ -70,6 +65,7 @@ Fixed.propTypes = {
   className: PropTypes.string,
   style: PropTypes.object,
   fit: PropTypes.bool,
+  children: PropTypes.any,
 };
 
 export default Fixed;